refactor: snapshotApi를 제외한 모든 클래스 패키지 이동 완료
