+++ src/main/java/com/movte/slate/domain/snapshot/application/service/response/SearchBunchOfSnapshotWithMovieTitleServiceResponse.java
@@ -1,4 +1,4 @@
-package com.movte.slate.domain.stillcut.application.service.response;
+package com.movte.slate.domain.snapshot.application.service.response;
 
 import com.movte.slate.domain.movie.application.service.dto.MovieResponseDto;
 import java.util.List;
