+++ src/main/java/com/movte/slate/domain/snapshot/domain/Snapshot.java
@@ -1,4 +1,4 @@
-package com.movte.slate.domain.stillcut.domain;
+package com.movte.slate.domain.snapshot.domain;
 
 import com.movte.slate.domain.snapshot.domain.Scene;
 import com.movte.slate.domain.user.domain.User;
