+++ src/main/java/com/movte/slate/domain/snapshot/repository/SnapShotRepository.java
@@ -1,6 +1,6 @@
 package com.movte.slate.domain.snapshot.repository;
 
-import com.movte.slate.domain.stillcut.domain.Snapshot;
+import com.movte.slate.domain.snapshot.domain.Snapshot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SnapShotRepository extends JpaRepository<Snapshot, Long> {
